Added gradient animation on Jumbotron
Added animation on App.css, declaring the properties.
Added declaration of background-size. The size needs to exceed 100%, as the animation relies on the translation on the position of the background gradient.
The background size and image gradient is set as important to prevent darkreader from changing it.
Condition-based background image declaration is added to Jumbotron. It uses a reference callback as React do not support setting of !important through the normal React method of setting the styles.