Removed OTP input from RegistrationForm in favor of only using only Google OAuth
This is done because it seems redundant to have both OAuth and OTP.
The OTP input is instead hidden. It will be used as a hidden session validator of sorts that the website will store a session code together with the email from OAuth. This would make sure that the email from the form input really came from Google OAuth.