Fix recurring schedules not showing beyond one month
Added an addDays function on Dates for DragAndDropCalendar to use.
The rrule now uses the `between` function to clip the excess dates now yet rendered.
The `between` function uses the pattern `new Date(new Date(date).toDateString())` to remove the time component on the date, as it seems to also remove dates that have the same day but is 'below' the 'clipping time'.